Update on "[JIT] add other module apis"
[ghstack-poisoned]
